pitron practice

+++ week 01 -- Introduction/Module 04 - Assignment 01 (Hackeerramk)/Get Prefix Sum.cpp
@@ -4,20 +4,16 @@ int main()
 {
     int n;
     cin >>n;
-    vector <int> ar(n);
+    vector <long long int> ar(n);
     for (int i=0; i<ar.size(); i++)
     {
         cin>>ar[i];
     }
-    vector <int> pre(n);
+    vector <long long int> pre(n);
     pre[0]=ar[0];
     for (int i=1; i<pre.size(); i++)
     {
         pre[i]=ar[i]+pre[i-1];
-      //  cout <<pre[i]<<" ";
- //   reverse(pre.begin(),pre.end());
-   // cout <<pre[i]<<" ";
-
 
     }
     reverse(pre.begin(),pre.end());
